Move metadata tag library actions into module
